Upgrade mathjax to 2.7.5
For included fixes see:

* https://www.mathjax.org/mathjax-v2-7-2-now-available/
* https://www.mathjax.org/mathjax-v2-7-3-now-available/
* https://www.mathjax.org/mathjax-v2-7-4-now-available/
* https://www.mathjax.org/mathjax-v2-7-5-now-available/

Version 2.7.4 also fixes CVE-2018-1999024, which is a potential
cross-site scripting vulnerability. I couldn't get the exploit to work
in Discourse so the existing sanitation probably did its job, but it's
always a risk that someone is clever enough to bypass that.

The public/mathjax directory is basically a copy of the corresponding
files in https://github.com/mathjax/MathJax/tree/2.7.5 but I included
only those font directories that existed in the tree already
(the HTML-CSS/TeX tree without the png subdirectory).